wikirecent: simplify, copyedit, and port to mzworkflows.py
Touches #9809. I also substantially pared down the amount of code
involved without losing any features. No need to have Python utility
scripts when `psql` and `curl` do the trick.
