+++ misc/perf-dash/app/Dockerfile
@@ -7,7 +7,7 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-MZFROM wikirecent-app
+MZFROM wikirecent-server
 
 # Override the wikirecent HTML / JS template files with our own
 COPY templates /app/templates
