--- play/wikirecent/app/src/templates/index.html
@@ -1,81 +1,0 @@
-<html>
-  <head>
-    <title>{% block title %}Wikipedia Live Edits!{% end %}</title>
-
-    <script src="https://cdn.jsdelivr.net/npm/vega@5"></script>
-    <script src="https://cdn.jsdelivr.net/npm/vega-lite@4"></script>
-    <script src="https://cdn.jsdelivr.net/npm/vega-embed@6"></script>
-
-  </head>
-  <body>
-      <div id="edit counts"><h1><span id="counter">0</span> Edits Made!</h1></div>
-      <div id="top10"></div>
-
-    <script type="text/javascript">
-        "use strict";
-
-        var path = "ws://" + location.host + "{{reverse_url('api/stream', 'counter')}}";
-        var connection = new WebSocket(path);
-
-        connection.onmessage = function(event) {
-                var data = JSON.parse(event.data);
-                // Counter is a single row table, so every update should contain one insert and
-                // maybe one delete (which we don't care about)
-                document.getElementById("counter").innerHTML = data.inserted[0][0]
-        }
-
-    </script>
-
-    <script type="text/javascript">
-        "use strict";
-
-        function setup_view(view_name) {
-
-            var chart = {
-                $schema: 'https://vega.github.io/schema/vega-lite/v4.json',
-                mark: 'bar',
-                height: 400,
-                width: 800,
-                data: {name: 'data'},
-                encoding: {
-                    x: {field: 'subject', type: 'nominal', sort: 'y'},
-                    y: {field: 'count', type: 'quantitative'},
-                }
-            };
-
-            var config = {
-                actions: false
-            }
-
-            vegaEmbed('#' + view_name, chart, config).then(function(chart) {
-
-                var path = "ws://" + location.host + "{{reverse_url('api/stream', '')}}" + view_name;
-                var connection = new WebSocket(path);
-
-                function convert_to_subject(row) {
-                    return {subject: row[0], count: parseInt(row[1])};
-                }
-
-                function subject_in_array(e, arr) {
-                    return arr.find(i => i.subject === e.subject && i.count === e.count);
-                }
-
-                connection.onmessage = function(event) {
-                        var data = JSON.parse(event.data);
-                        var insert_values = data.inserted.map(convert_to_subject);
-                        var delete_values = data.deleted.map(convert_to_subject);
-                        var changeSet = vega.changeset()
-                                .insert(insert_values)
-                                .remove(d => subject_in_array(d, delete_values));
-
-                        chart.view.change('data', changeSet).resize().run();
-                }
-            });
-
-        }
-
-        setup_view("top10");
-
-    </script>
-  </body>
-</html>
