+++ play/wikirecent/server/Dockerfile
@@ -1,4 +1,3 @@
-#!/usr/bin/env bash
 # Copyright Materialize, Inc. and contributors. All rights reserved.
 #
 # Use of this software is governed by the Business Source License
@@ -8,4 +7,13 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-psql -U materialize -h materialized -p 6875 -f /data/views.sql materialize
+MZFROM ubuntu-base
+
+RUN apt-get update && apt-get install -y libpq5 python3 python3-pip
+
+COPY requirements.txt .
+RUN pip install -r requirements.txt && rm requirements.txt
+
+WORKDIR /server
+COPY index.html serve.py ./
+ENTRYPOINT ["python3", "serve.py"]
