+++ play/wikirecent/server/mzbuild.yml
@@ -7,10 +7,4 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-FROM postgres
-
-RUN mkdir /data
-COPY views.sql /data
-
-COPY create_views /usr/local/bin
-CMD /usr/local/bin/create_views
+name: wikirecent-server
