+++ play/wikirecent/server/requirements.txt
@@ -1,0 +1,2 @@
+psycopg==3.0.7
+tornado==6.1
